requirements.txt
openai>=1.0.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
scipy>=1.11.0

